chore(lint): apply 4607 ruff auto-fixes across 1343 files
Applied safe ruff auto-fixes including:
- Removed unused imports (F401)
- Fixed import ordering (I001)
- Simplified nested if statements (SIM102)
- Other safe automatic fixes

Impact:
- 1343 files modified
- 4885 insertions, 5677 deletions (net -792 lines)
- 4607 linting errors fixed automatically
- 4239 errors remain (require manual review)

Ruff fixes applied while Copilot investigates remaining smoke test issues (199 collection errors).

+++ scripts/phase4_build_canary.py
@@ -62,7 +62,7 @@ def main():
         by_top[top].append((key, str(m)))
 
     total = sum(len(v) for v in by_top.values())
-    target_total = max(1, int(round(total * args.size)))
+    target_total = max(1, round(total * args.size))
 
     # proportional per top-level, deterministic by hash
     picks = []
@@ -70,7 +70,7 @@ def main():
     for top, items in by_top.items():
         if top not in TOPS:
             continue
-        need = int(round(len(items) / total * target_total))
+        need = round(len(items) / total * target_total)
         items_sorted = sorted(items, key=lambda kv: sha_bucket(kv[0]))
         picks.extend(items_sorted[:need])
         allocated += need
